anonyfiles_cli/requirements.txt
spacy>=3.5.2,<4.0.0
pandas>=1.5.0
numpy>=1.23.5
thinc>=8.1.8,<8.3.0
python-docx>=0.8.11
faker>=18.0.0
# pykwalify  <-- peut être retiré si tu passes totalement à yamale
https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.5.0/fr_core_news_sm-3.5.0.tar.gz#egg=fr_core_news_sm
https://github.com/explosion/spacy-models/releases/download/fr_core_news_md-3.5.0/fr_core_news_md-3.5.0.tar.gz#egg=fr_core_news_md
openpyxl
PyYAML>=6.0
pymupdf
typer
yamale
docx

